refactor: Refactor project structure
This commit does not include refactoring the Components folder
