+++ package.json
@@ -31,15 +31,15 @@
     "analyze": "source-map-explorer 'build/static/js/*.js' --html bundle-analysis.html",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "palvelukartta-fetch-islands": "node src/scripts/fetchPalvelukarttaIslandData.js",
-    "palvelukartta-append-island-data": "node sr/scripts/mergePalvelukarttaIslandDataAdditions.js"
+    "palvelukartta-fetch-islands": "node src/data/scripts/fetchPalvelukarttaIslandData.js",
+    "palvelukartta-append-island-data": "node src/data/scripts/mergePalvelukarttaIslandDataAdditions.js"
   },
   "eslintConfig": {
     "extends": "react-app"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged && pretty-quick --staged && node src/scripts/mapDataManager.js",
+      "pre-commit": "lint-staged && pretty-quick --staged && node src/data/scripts/mapDataManager.js",
       "prepare-commit-msg": "exec < /dev/tty && git cz --hook"
     }
   },
