+++ src/Components/DropdownMenu/DropdownMenu.js
@@ -6,7 +6,7 @@ import BaseButton from '../BaseButton/BaseButton';
 import UnstyledLink from '../UnstyledLink/UnstyledLink';
 import { ReactComponent as Menu } from '../../assets/icons/menu.svg';
 import { ReactComponent as Exit } from '../../assets/icons/exit.svg';
-import { POINT_TYPES } from '../../App';
+import { POINT_TYPES } from '../../domain/app/App';
 
 import styled, { keyframes } from 'styled-components';
