+++ src/assets/fonts/HelsinkiGrotesk/HelsinkiGrotesk-RegularItalic.otf
