+++ src/assets/styles/fonts.css
@@ -1,13 +1,13 @@
 @font-face {
   font-family: 'Helsinki Grotesk Regular';
-  src: url('../assets/HelsinkiGrotesk/HelsinkiGrotesk-Regular.otf')
+  src: url('../fonts/HelsinkiGrotesk/HelsinkiGrotesk-Regular.otf')
     format('opentype');
   font-weight: 400;
 }
 
 @font-face {
   font-family: 'Helsinki Grotesk Bold';
-  src: url('../assets/HelsinkiGrotesk/HelsinkiGrotesk-Bold.otf')
+  src: url('../fonts/HelsinkiGrotesk/HelsinkiGrotesk-Bold.otf')
     format('opentype');
   font-weight: 700;
 }
