+++ src/assets/styles/index.scss
@@ -1,7 +1,7 @@
-@import '~sanitize.css';
+@import '../../../node_modules/sanitize.css';
 @import './html5doctor.css';
 @import './fonts.css';
-@import './slick.scss';
+@import 'slick';
 
 /*variables*/
