+++ src/assets/styles/slick.scss
@@ -1,6 +1,6 @@
 // Slick carousel base CSS
-@import '~slick-carousel/slick/slick.css';
-@import '~slick-carousel/slick/slick-theme.css';
+@import '../../../node_modules/slick-carousel/slick/slick.css';
+@import '../../../node_modules/slick-carousel/slick/slick-theme.css';
 
 // Style the buttons to be more obvious
 .slick-prev,
@@ -59,19 +59,19 @@
 
 .slick-prev::before {
   transform: rotate(180deg);
-  content: url('../assets/icons/chevron_right.svg');
+  content: url('../icons/chevron_right.svg');
 }
 
 [dir='rtl'] .slick-prev::before {
   transform: rotate(0deg);
-  content: url('../assets/icons/chevron_right.svg');
+  content: url('../icons/chevron_right.svg');
 }
 
 .slick-next::before {
-  content: url('../assets/icons/chevron_right.svg');
+  content: url('../icons/chevron_right.svg');
 }
 
 [dir='rtl'] .slick-next::before {
   transform: rotate(180deg);
-  content: url('../assets/icons/chevron_right.svg');
+  content: url('../icons/chevron_right.svg');
 }
