+++ src/data/scripts/mapDataManager.js
