+++ src/domain/app/App.js
@@ -1,16 +1,16 @@
 import React, { Suspense } from 'react';
 import CookieConsent from 'react-cookie-consent';
-import MapPage from './Components/MapPage/MapPage';
+import MapPage from '../../Components/MapPage/MapPage';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import Home from './Components/Home/Home';
+import Home from '../../Components/Home/Home';
 import { withTranslation } from 'react-i18next';
-import ApplicationHeader from './Components/ApplicationHeader/ApplicationHeader';
-import mapData from './mapData.json';
-import lineData from './lineData.json';
-import mapIslandData from './mapIslandData.json';
+import ApplicationHeader from '../../Components/ApplicationHeader/ApplicationHeader';
+import mapData from '../../data/mapData.json';
+import lineData from '../../data/lineData.json';
+import mapIslandData from '../../data/mapIslandData.json';
 import { ThemeProvider } from 'styled-components';
-import withTracker from './withTracker';
-import ScrollToTop from './ScrollToTop';
+import withTracker from '../../common/utils/withTracker';
+import ScrollToTop from '../../common/utils/ScrollToTop';
 export const GlobalGeoContext = React.createContext();
 export const GlobalLineContext = React.createContext();
 export const GlobalIslandContext = React.createContext();
